requirements.txt
skan
rdp
pulp
numpy
opencv-python
pulp
PyYAML
torch
torchvision
matplotlib